@@ -2,7 +2,8 @@
   "name": "dir-contents-object",
   "version": "1.0.0",
   "description": "Convert the contents of a directory to an object",
-  "main": "dist/main.js",
+  "type": "module",
+  "exports": "./dist/main.js",
   "types": "dist/main.d.ts",
   "author": "Mgen (https://www.mgenware.com)",
   "homepage": "https://github.com/mgenware/dir-contents-object",
@@ -18,10 +19,10 @@
   "license": "MIT",
   "scripts": {
     "r": "daizong",
-    "prepublishOnly": "yarn r build"
+    "test": "daizong build"
   },
   "engines": {
-    "node": ">=10.0.0"
+    "node": ">=12.0.0"
   },
   "files": [
     "/dist/**/*.js",
@@ -30,19 +31,20 @@
     "/dist/**/*.map"
   ],
   "devDependencies": {
-    "@types/mocha": "^7.0.2",
-    "@types/node": "^14.0.19",
-    "@typescript-eslint/eslint-plugin": "^3.6.0",
-    "@typescript-eslint/parser": "^3.6.0",
-    "daizong": "^0.7.0",
-    "eslint": "^7.4.0",
-    "eslint-config-airbnb-typescript-lite": "^9.0.0",
-    "eslint-plugin-import": "^2.22.0",
-    "mocha": "^8.0.1",
-    "source-map-support": "^0.5.19",
-    "typescript": "^3.9.6"
+    "@types/mocha": "^9.0.0",
+    "@types/node": "^17.0.5",
+    "@typescript-eslint/eslint-plugin": "^5.8.0",
+    "@typescript-eslint/parser": "^5.8.0",
+    "daizong": "^0.22.0",
+    "eslint": "^8.5.0",
+    "eslint-config-mgenware": "^0.6.1",
+    "eslint-plugin-import": "^2.25.3",
+    "mgenware-tsconfig-node": "^14.0.2",
+    "mocha": "^9.1.3",
+    "source-map-support": "^0.5.21",
+    "typescript": "^4.5.4"
   },
   "dependencies": {
-    "m-fs": "^4.0.0"
+    "m-fs": "^5.1.0"
   }
 }
