@@ -1,5 +1,5 @@
 import * as assert from 'assert';
-import getObject from '..';
+import getObject from '../dist/main.js';
 
 it('Main', async () => {
   assert.deepEqual(await getObject('./tests/data'), {
